Extract WebSocketMessage into its own file
